+++ apps/api/src/app/environments/environments.controller.ts
@@ -19,7 +19,7 @@ import { UpdateEnvironment } from './usecases/update-environment/update-environm
 import { UpdateEnvironmentRequestDto } from './dtos/update-environment-request.dto';
 import { ApiCommonResponses, ApiResponse } from '../shared/framework/response.decorator';
 import { UserAuthentication } from '../shared/framework/swagger/api.key.security';
-import { SdkGroupName, SdkMethodName } from '../shared/framework/swagger/method.name.decorator';
+import { SdkGroupName, SdkMethodName } from '../shared/framework/swagger/sdk.decorators';
 
 @ApiCommonResponses()
 @Controller('/environments')
