+++ apps/api/src/app/subscribers/subscribers.controller.ts
@@ -101,9 +101,7 @@ import { MarkMessageAsByMarkCommand } from '../widgets/usecases/mark-message-as-
 import { MarkMessageAsByMark } from '../widgets/usecases/mark-message-as-by-mark/mark-message-as-by-mark.usecase';
 import { FeedResponseDto } from '../widgets/dtos/feeds-response.dto';
 import { UserAuthentication } from '../shared/framework/swagger/api.key.security';
-import { SdkGroupName, SdkMethodName } from '../shared/framework/swagger/method.name.decorator';
-
-import { SdkUsePagination } from '../shared/framework/swagger/sdk.decorators';
+import { SdkGroupName, SdkMethodName, SdkUsePagination } from '../shared/framework/swagger/sdk.decorators';
 
 @ThrottlerCategory(ApiRateLimitCategoryEnum.CONFIGURATION)
 @ApiCommonResponses()
