+++ apps/api/src/app/workflows/workflow.controller.ts
@@ -47,7 +47,7 @@ import { DeleteNotificationTemplateCommand } from './usecases/delete-notificatio
 import { GetWorkflowVariables } from './usecases/get-workflow-variables/get-workflow-variables.usecase';
 import { GetWorkflowVariablesCommand } from './usecases/get-workflow-variables/get-workflow-variables.command';
 import { UserAuthentication } from '../shared/framework/swagger/api.key.security';
-import { SdkGroupName } from '../shared/framework/swagger/method.name.decorator';
+import { SdkGroupName } from '../shared/framework/swagger/sdk.decorators';
 
 @ApiCommonResponses()
 @Controller('/workflows')
